Implement Contents class
This almost acts like an array, and as such can be used as drop-in
replacement for `$c` – well, we need to make a couple of adjustments
in the core, and likely some plugins would also need some adjustments.
This class can be adapted to work for other markup languages, but
Markdown has the advantage that it's simple and basically an extension
of HTML, so even existing HTML contents work out of the box.  And of
course, the class can be modified to work with HTML.